fronted_Juana/requirements.txt
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
pycodestyle==2.7.0
toml==0.10.2
Werkzeug==2.0.1
WTForms==2.3.3
python-dotenv==0.20.0